disable language picker for Android 13+
This is better dealt with from the system settings in this case, so
direct the user there.

Otherwise it would get quite messy with synchronizing the selected
language from the system back into our own preference. That's probably
doable somehow but it seems entirely not worth it with dealing with
language code fallback chains for this scenario where a user can just
use the system menu now.