don't try to handle orientation changes ourself
This led to broken toolbar layout when the app was started in a
different orientation than the current one.

This might have some adverse effects now as we trigger an activity
restart on every config change which might not be handled gracefully in
all cases. So far all seems to work though and we can deal with any
potential breakage later.

Fixes: #118 #113 (potentially) #117 (potentially)